Fix remove-older-than command
Move the command to the same line as the backup command. This was broken
by the quotes around the backup URL. But these quotes are useful
because URLs may contain special shell characters. Move the command to
the same line in the template to avoid that someone overlooks the clever
use of "-%>" in the template.

Also refines the test to detect this in the future.